Moved the compatibility logic into multiport_adapter.go
