+++ agent/consul/prepared_query_endpoint.go
@@ -15,6 +15,7 @@ import (
 	"github.com/hashicorp/go-uuid"
 
 	"github.com/hashicorp/consul/acl"
+	"github.com/hashicorp/consul/agent/consul/adapter"
 	"github.com/hashicorp/consul/agent/consul/state"
 	"github.com/hashicorp/consul/agent/structs"
 	"github.com/hashicorp/consul/agent/structs/aclfilter"
@@ -489,7 +490,7 @@ func (p *PreparedQuery) Execute(args *structs.PreparedQueryExecuteRequest,
 		}
 	}
 
-	populateLegacyCheckServiceNodePorts(reply.Nodes)
+	adapter.PopulateLegacyCheckServiceNodePorts(reply.Nodes)
 	return nil
 }
 
@@ -540,7 +541,7 @@ func (p *PreparedQuery) ExecuteRemote(args *structs.PreparedQueryExecuteRemoteRe
 		reply.Nodes = reply.Nodes[:args.Limit]
 	}
 
-	populateLegacyCheckServiceNodePorts(reply.Nodes)
+	adapter.PopulateLegacyCheckServiceNodePorts(reply.Nodes)
 	return nil
 }
